check directory and file
- check whether directory and file exist
- delete cat _vimrc > .vimrc

+++ install.sh
@@ -6,10 +6,15 @@ echo "all your old vim settings will be mv to .vimrc.old & .vim.old"
 echo -en "Do you want to continue?[y/n]"
 read -n 1 use_vim_configs
 if [ "$use_vim_configs" == "Y" ] || [ "$use_vim_configs" == "y" ]; then
-	mv $VIMDIR/.vim $VIMDIR/.vim.old
-	mv $VIMDIR/.vimrc $VIMDIR/.vimrc.old
+	if [ -d "$VIMDIR/.vim" ]; then
+		mv $VIMDIR/.vim $VIMDIR/.vim.old
+		echo "origin .vim directory has been moved to .vim.old"
+	fi
+	if [ -f "$VIMDIR/.vimrc" ]; then
+		mv $VIMDIR/.vimrc $VIMDIR/.vimrc.old
+		echo "origin .vimrc has been moved to .vimrc.old"
+	fi
 	ln -s $PWD/.vim $VIMDIR/.vim 2> /dev/null
-	cat $PWD/_vimrc > $PWD/.vimrc
 	ln -s $PWD/.vimrc $VIMDIR/.vimrc 2> /dev/null
 fi
